🚚 Rename: 카드 도메인 컴포넌트를 features 폴더로 이동(ActivityCard, MyActivityCard, ResevationCard)
