+++ src/features/main/ui/activity-card/ActivityCard.tsx
